requirements.txt
Django
Pillow
openpyxl
python-dotenv
django-simple-captcha
html-sanitizer
djangorestframework
djangorestframework-simplejwt
pytest
pytest-django
pytest-freezegun
freezegun
